moved http client creation to new file to share between controllers
