+++ controllers/client.go
@@ -1,0 +1,19 @@
+package controllers
+
+import (
+	"crypto/tls"
+	"github.com/RedHatInsights/entitlements-api-go/config"
+	"net/http"
+)
+
+func getClient() *http.Client {
+	// Create a HTTPS client that uses the supplied pub/priv mutual TLS certs
+	return &http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{
+				RootCAs:      config.GetConfig().RootCAs,
+				Certificates: []tls.Certificate{*config.GetConfig().Certs},
+			},
+		},
+	}
+}
